fix cuda graphs
