+++ aiter/ops/triton/_triton_kernels/flash_attn_triton_amd/fwd_decode.py
@@ -966,6 +966,7 @@ def attention_forward_decode_triton_impl(
     rotary_sin: Optional[torch.Tensor] = None,
     rotary_interleaved: bool = False,
     seqlens_rotary: Optional[torch.Tensor] = None,
+    max_seqlen_k: Optional[int] = None,
 ):
     # Validate layout at entry
     if layout != "bshd":
@@ -1066,11 +1067,13 @@ def attention_forward_decode_triton_impl(
         # For paged attention, k_cache and v_cache have shape [num_blocks, block_size, nheads, head_dim]
         num_blocks_kc, block_size_k, nheads_kc, dim_kc = k_cache.shape
         num_blocks_vc, block_size_v, nheads_vc, dim_vc = v_cache.shape
-        # Get the actual sequence length from cache_seqlens or block_table
-        if cache_seqlens is not None:
-            seqlen_kc = int(cache_seqlens.max().item())
+        # Get the actual sequence length from cache_seqlens or block_table.
+        # Avoid cache_seqlens.max().item() — GPU-to-CPU sync is illegal during
+        # CUDA graph capture.  Use max_seqlen_k when provided (tightest bound),
+        # otherwise fall back to the block_table upper bound which is always safe.
+        if max_seqlen_k is not None:
+            seqlen_kc = max_seqlen_k
         else:
-            # Infer from block_table shape [batch_size, num_blocks_per_seq]
             assert block_table is not None
             num_blocks_per_seq = block_table.shape[1]
             seqlen_kc = num_blocks_per_seq * block_size_k
@@ -1141,13 +1144,8 @@ def attention_forward_decode_triton_impl(
 
     # Use heuristics for split_k
     if use_block_table:
-        # For paged attention, use the actual sequence length from cache_seqlens
-        max_seqlen = (
-            int(cache_seqlens.max().item())
-            if cache_seqlens is not None
-            else block_size_k
-        )
-        split_k = get_split_k(batch_size, n_group_q, heads_per_group_q, max_seqlen)
+        # Reuse seqlen_kc (already computed above without GPU-to-CPU sync)
+        split_k = get_split_k(batch_size, n_group_q, heads_per_group_q, seqlen_kc)
     else:
         split_k = get_split_k(batch_size, n_group_q, heads_per_group_q, seqlen_kc)
     split_size = (seqlen_kc + split_k - 1) // split_k
