+++ aiter/ops/triton/_triton_kernels/flash_attn_triton_amd/interface_v3.py
@@ -290,6 +290,7 @@ def fwd(
             rotary_sin=rotary_sin,
             rotary_interleaved=rotary_interleaved,
             seqlens_rotary=seqlens_rotary,
+            max_seqlen_k=max_seqlen_k,
         )
     else:
         if DEBUG:
